Move ScaleData to Ds namespace
